src/test/java/br/com/munieri/desafio/questao9/BinaryTreeTest.java
package br.com.munieri.desafio.questao9;

import org.junit.Test;

import static org.junit.Assert.*;

public class BinaryTreeTest {

    @Test
    public void deve_verificar_existencia_elementos() {
        BinaryTree bt = createBinaryTree();

        assertTrue(bt.containsNode(6));
        assertTrue(bt.containsNode(4));

        assertFalse(bt.containsNode(1));
    }

    @Test
    public void deve_retornar_soma() {

        BinaryTree bt = createBinaryTree();

        assertEquals(bt.sum(), 42);
    }

    private BinaryTree createBinaryTree() {
        BinaryTree bt = new BinaryTree();

        bt.add(6);
        bt.add(4);
        bt.add(8);
        bt.add(3);
        bt.add(5);
        bt.add(7);
        bt.add(9);

        return bt;
    }

}